resolve các comment
