+++ front-end/src/app/tin-tuc-su-kien/tin-tuc/[article]/content.tsx
@@ -1,17 +1,12 @@
 
-export default function ArticleContent({content, describe}:{content:string, describe:string}){
-
-    const formatContent = ()=> content.split('\n').map((line, idx) => (
-        <span key={idx}>
-            {line}
-            <br />
-        </span>
-    ))
+export default function ArticleContent({ content, describe }: { content: string, describe: string }) {
 
     return (
         <div className="py-10 md:px-20 w-3/4">
             <p className="mt-20 font-bold">{describe}</p>
-            <p className="mt-5">{formatContent()}</p>
-        </div>
+            <pre className="whitespace-pre-wrap">
+                <span dangerouslySetInnerHTML={{ __html: content }} />
+            </pre>
+        </div >
     )
 }
