+++ front-end/src/components/lastest-news-list-symbol.tsx
@@ -1,0 +1,17 @@
+
+export default function LatestNewsListSymbol() {
+    return (
+        <svg className="h-10 w-10 text-gray-400 md-5 mr-2"
+            viewBox="0 0 24 24"
+            style={{ flexShrink: 0 }}
+            stroke-width="1"
+            stroke="currentColor"
+            fill="none"
+            stroke-linecap="round"
+            stroke-linejoin="round">
+            <path stroke="none" d="M0 0h24v24H0z" />
+            <line x1="12" y1="5" x2="12" y2="19" />
+            <line x1="5" y1="12" x2="19" y2="12" />
+        </svg>
+    )
+}
